+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @elf-framework/design-system
 
+## 2.1.41
+
+### Patch Changes
+
+- support Animation component
+  modify button, toolbar style
+  support table component
+- Updated dependencies []:
+  - @elf-framework/design-tokens@2.1.41
+
 ## 2.1.40
 
 ### Patch Changes